fix: inveters auto resend, changed power map, separated brake implausibility thr, steer angle

+++ Core/Inc/inverters.h
@@ -30,7 +30,7 @@ typedef enum {
 #define INV_CUTOFF_COEFF_REAL       INV_POWER_LIMIT*60/(MOT_TORQUE_COEFF*2*M_PI) // A/s
 #define INV_CURR_PEAK_REAL          169.9f // A
 #define INV_RPM_CUTOFF              INV_CUTOFF_COEFF_REAL/INV_CURR_PEAK_REAL+1 // 1/s
-#define INV_CUTOFF_COEFF            (int32_t)(INV_CUTOFF_COEFF_REAL/INV_CURR_PEAK_REAL*INT16_MAX) // 1/s
+#define INV_CUTOFF_COEFF            (INV_CUTOFF_COEFF_REAL/INV_CURR_PEAK_REAL*INT16_MAX) // 1/s
 #define INV_CUTOFF_COEFF_TORQUE     MOT_TORQUE_COEFF / INV_CUTOFF_COEFF_REAL // 1/s
 
 void INV_enable_regid_updates(uint16_t regid, uint8_t interval);
